traffic-backend/requirements.txt
Flask==2.2.5
Flask-Cors==3.0.10
numpy
opencv-python
torch
torchvision
twilio
paho-mqtt
RPi.GPIO
